add synced config toggle for AzuEPI Backpack slot
We add a new synced configuration toggle allowing to enable or disable
adding the Backpack slot to AzuExtendedPlayerInventory (defaults to
enabled).

With this, it is possible to have the Backpacks mod installed yet hide
it completely from the users by disabling the default Explorers Backpack
and removing the AzuExtendedPlayerInventory slot. This can be useful
when the Backpacks mod is used as a dependency for something else but
allowing backpacks themselves is not desired.

For example, BowsBeforeHoes[1] depends on Backpacks to implement its
quiver functionallity, however BowsBeforeHoes users might want to only
enable quivers and not backpacks.

[1]: https://thunderstore.io/c/valheim/p/Azumatt/BowsBeforeHoes/